src/auth-service/requirements.txt
psycopg2==2.9.11
fastapi==0.120.2
python-dotenv==1.2.1
uvicorn==0.38.0